--- src/main/java/com/kosa/ownit/OwnItApplication.java
@@ -1,13 +1,0 @@
-package com.kosa.ownit;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OwnItApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OwnItApplication.class, args);
-    }
-
-}
