--- src/main/java/com/kosa/ownit/controller/car/CarAdminController.java
@@ -1,10 +1,0 @@
-package com.kosa.ownit.controller.car;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/admin")
-public class CarAdminController {
-
-}
