--- src/main/java/com/kosa/ownit/controller/car/CarUserController.java
@@ -1,11 +1,0 @@
-package com.kosa.ownit.controller.car;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/user")
-public class CarUserController {
-
-
-}
