--- src/main/java/com/kosa/ownit/controller/mr/MrAdminController.java
@@ -1,7 +1,0 @@
-package com.kosa.ownit.controller.mr;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class MrAdminController {
-}
