--- src/main/java/com/kosa/ownit/domain/car/CarDTO.java
@@ -1,4 +1,0 @@
-package com.kosa.ownit.domain.car;
-
-public class CarDTO {
-}
