--- src/main/java/com/kosa/ownit/domain/car/CarVO.java
@@ -1,4 +1,0 @@
-package com.kosa.ownit.domain.car;
-
-public class CarVO {
-}
