--- src/main/java/com/kosa/ownit/domain/mr/MrDTO.java
@@ -1,4 +1,0 @@
-package com.kosa.ownit.domain.mr;
-
-public class MrDTO {
-}
