--- src/main/java/com/kosa/ownit/domain/mr/MrVO.java
@@ -1,4 +1,0 @@
-package com.kosa.ownit.domain.mr;
-
-public class MrVO {
-}
