--- src/main/java/com/kosa/ownit/mapper/car/CarAdminMapper.java
@@ -1,10 +1,0 @@
-package com.kosa.ownit.mapper.car;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import java.util.Date;
-
-@Mapper
-public interface CarAdminMapper {
-    public Date test();
-}
