--- src/main/java/com/kosa/ownit/mapper/car/CarUserMapper.java
@@ -1,8 +1,0 @@
-package com.kosa.ownit.mapper.car;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface CarUserMapper {
-
-}
