--- src/main/java/com/kosa/ownit/mapper/mr/MrAdminMapper.java
@@ -1,4 +1,0 @@
-package com.kosa.ownit.mapper.mr;
-
-public interface MrAdminMapper {
-}
