--- src/main/java/com/kosa/ownit/service/car/CarUserService.java
@@ -1,5 +1,0 @@
-package com.kosa.ownit.service.car;
-
-
-public interface CarUserService {
-}
