--- src/main/java/com/kosa/ownit/service/car/CarUserServiceImpl.java
@@ -1,9 +1,0 @@
-package com.kosa.ownit.service.car;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class CarUserServiceImpl implements CarUserService{
-
-}
