--- src/main/java/com/kosa/ownit/service/mr/MrUserService.java
@@ -1,4 +1,0 @@
-package com.kosa.ownit.service.mr;
-
-public interface MrUserService {
-}
