--- src/main/java/com/kosa/ownit/service/mr/MrUserServiceImpl.java
@@ -1,7 +1,0 @@
-package com.kosa.ownit.service.mr;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class MrUserServiceImpl implements MrUserService{
-}
