+++ src/main/java/com/kosa/resq/ResQApplication.java
@@ -1,0 +1,13 @@
+package com.kosa.resq;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ResQApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ResQApplication.class, args);
+    }
+
+}
