+++ src/main/java/com/kosa/resq/controller/car/CarAdminController.java
@@ -1,0 +1,10 @@
+package com.kosa.resq.controller.car;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/admin")
+public class CarAdminController {
+
+}
