+++ src/main/java/com/kosa/resq/controller/car/CarUserController.java
@@ -1,0 +1,11 @@
+package com.kosa.resq.controller.car;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/user")
+public class CarUserController {
+
+
+}
