+++ src/main/java/com/kosa/resq/controller/mr/MrAdminController.java
@@ -1,0 +1,7 @@
+package com.kosa.resq.controller.mr;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MrAdminController {
+}
