+++ src/main/java/com/kosa/resq/controller/mr/MrUserController.java
@@ -1,0 +1,7 @@
+package com.kosa.resq.controller.mr;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MrUserController {
+}
