+++ src/main/java/com/kosa/resq/domain/car/CarVO.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.domain.car;
+
+public class CarVO {
+}
