+++ src/main/java/com/kosa/resq/domain/mr/MrVO.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.domain.mr;
+
+public class MrVO {
+}
