+++ src/main/java/com/kosa/resq/mapper/car/CarAdminMapper.java
@@ -1,0 +1,10 @@
+package com.kosa.resq.mapper.car;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import java.util.Date;
+
+@Mapper
+public interface CarAdminMapper {
+    public Date test();
+}
