+++ src/main/java/com/kosa/resq/mapper/car/CarUserMapper.java
@@ -1,0 +1,8 @@
+package com.kosa.resq.mapper.car;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface CarUserMapper {
+
+}
