+++ src/main/java/com/kosa/resq/mapper/mr/MrAdminMapper.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.mapper.mr;
+
+public interface MrAdminMapper {
+}
