+++ src/main/java/com/kosa/resq/service/car/CarUserServiceImpl.java
@@ -1,0 +1,8 @@
+package com.kosa.resq.service.car;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CarUserServiceImpl implements CarUserService{
+
+}
