+++ src/main/java/com/kosa/resq/service/mr/MrUserService.java
@@ -1,0 +1,4 @@
+package com.kosa.resq.service.mr;
+
+public interface MrUserService {
+}
