+++ src/main/java/com/kosa/resq/service/mr/MrUserServiceImpl.java
@@ -1,0 +1,7 @@
+package com.kosa.resq.service.mr;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class MrUserServiceImpl implements MrUserService{
+}
