+++ src/main/resources/mapper/car/CarAdminMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.kosa.ownit.mapper.car.CarAdminMapper">
+<mapper namespace="com.kosa.resq.mapper.car.CarAdminMapper">
     <select id="test">
         select sysdate
         from dual
