+++ src/main/resources/mapper/mr/MrAdminMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.kosa.ownit.mapper.mr.MrAdminMapper">
+<mapper namespace="com.kosa.resq.mapper.mr.MrAdminMapper">
     <select id="test">
         select sysdate
         from dual
