--- src/test/java/com/kosa/ownit/OwnItApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kosa.ownit;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OwnItApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
