+++ src/test/java/com/kosa/resq/ResQApplicationTests.java
@@ -1,0 +1,13 @@
+package com.kosa.resq;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ResQApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
