+++ src/test/java/com/kosa/resq/mapper/car/CarAdminMapperTest.java
@@ -1,0 +1,20 @@
+package com.kosa.resq.mapper.car;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
+
+
+@SpringBootTest
+@TestPropertySource(locations = "classpath:application.yaml")
+public class CarAdminMapperTest {
+    @Autowired
+    private CarAdminMapper mapper;
+
+    @Test
+    public void test(){
+        System.out.println(mapper.test());
+    }
+
+}
